Generalising Wordpress ViewModel
