+++ app/src/main/java/net/informatikag/thomapp/utils/handlers/ThomsLineRecyclerAdapter.kt
@@ -5,10 +5,10 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import net.informatikag.thomapp.MainActivity
 import net.informatikag.thomapp.R
+import net.informatikag.thomapp.utils.models.view.ThomsLineViewModel
 import net.informatikag.thomapp.viewables.fragments.ThomsLine.main.ThomsLineFragment
 import net.informatikag.thomapp.viewables.viewholders.ThomsLineArticleViewHolder
 import net.informatikag.thomapp.viewables.viewholders.ThomsLineLoadingViewholder
-import net.informatikag.thomapp.utils.models.view.ThomsLineViewModel
 import net.informatikag.thomapp.viewables.viewholders.ThomsLineEndViewholder
 
 /**
