+++ app/src/main/java/net/informatikag/thomapp/utils/models/view/ThomaeumViewModel.kt
@@ -1,0 +1,8 @@
+package net.informatikag.thomapp.utils.models.view
+
+import android.app.Application
+import net.informatikag.thomapp.MainActivity
+
+class ThomaeumViewModel(application: Application):WordpressViewModel(application) {
+    override val BASE_URL: String = MainActivity.THOMAEUM_BASE_URL
+}
