+++ app/src/main/java/net/informatikag/thomapp/viewables/fragments/ThomsLine/article/ThomsLineArticleFragment.kt
@@ -64,7 +64,7 @@ class ThomsLineArticleFragment : Fragment() {
             }
         }
         else {
-            article = WordpressArticle(args.id, false, this.requireContext())
+            article = WordpressArticle(args.id, false, viewmodel.BASE_URL, this.requireContext())
             { article, error -> articleRefreshCallback(article, error) }
             binding.thomslineArticleSwipeRefreshLayout.isRefreshing = true
         }
